Moved WIP function cells_from_line from wrapper API to class RHEALPixDGGS
